+++ internal/provider/provider.go
@@ -113,6 +113,7 @@ import (
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/loggingpapertrail"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/loggings3"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/loggingscalyr"
+	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/loggingsftp"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/loggingsplunk"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/loggingsumologic"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/loggingsyslog"
@@ -269,6 +270,7 @@ func (p *fastlyProvider) Resources(_ context.Context) []func() resource.Resource
 		loggingpapertrail.NewResource,
 		loggings3.NewResource,
 		loggingscalyr.NewResource,
+		loggingsftp.NewResource,
 		loggingsplunk.NewResource,
 		loggingsumologic.NewResource,
 		loggingsyslog.NewResource,
@@ -423,6 +425,7 @@ func (p *fastlyProvider) ListResources(_ context.Context) []func() list.ListReso
 		loggingpapertrail.NewListResource,
 		loggings3.NewListResource,
 		loggingscalyr.NewListResource,
+		loggingsftp.NewListResource,
 		loggingsplunk.NewListResource,
 		loggingsumologic.NewListResource,
 		loggingsyslog.NewListResource,